pyproject.toml
[project]
name = "study-cards"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "anthropic>=0.72.0",
    "fastapi>=0.120.2",
    "openai>=2.6.1",
    "psycopg2-binary>=2.9.7",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
